Extract split layout component, create vertical split
